refactor(client): shared ChatMessageBody for channel and DM messages
The ~450 near-identical lines of message-body rendering (content,
attachments, inline media, link previews, reactions) merge into one
chat-primitives component; the channel renderer becomes a thin adapter
and the DM copies are deleted. The shared body takes no client and no
ambient hooks — callers inject the id-space, federated file domain,
and mutations.

Unifying on the channel pipeline fixes real DM divergences: legacy
HTML is now DOMPurify-sanitized (the DM path parsed raw HTML), the
edited tag keys on the edited flag instead of updatedAt (which also
flips on pin/unpin), emoji-only messages render at emoji size,
non-media attachments render as file cards with the blob: download
name, and each DM row gets the channel's per-message error boundary.
